+++ src/device/function/cuda/CUDAFunction.h
@@ -25,7 +25,7 @@ public:
     void embedding(float* y, const int* x, const float* W, const int d, const int x_size) override;
     void masked_attention(float* y, float* q, float* k, float* v, float* scores, int* pos, int dim, int head_num, int seq_q, int seq_kv) override;
     void elem_multiply(float* y, const float* x1, const float* x2, const int size) override;
-    void max_index(float* index, float* x, const int n, const int num) override;
+    void max_index(int* index, float* x, const int n, const int num) override;
     void repeat_kv(float* o, float* in, int dim, int rep, int n) override;
 
 };
